Add in_weight to attack
Previously in the function 'attack', in_edges were given a weight
of 1 and side_edges' weight could be specified as a parameter.
Now in_edges' weight can also be specified as a parameter.
Also, it's possible to specify a weight of 0, which means
that the other kind of edges will be ignored.